+++ backend/canisters/cycles_dispenser/api/src/updates/mod.rs
@@ -1,4 +1,3 @@
 pub mod add_canister;
 pub mod c2c_request_cycles;
-pub mod set_governance_principals;
 pub mod update_config;