+++ backend/canisters/proposals_bot/api/src/updates/remove_governance_canister.rs
@@ -1,4 +1,5 @@
 use candid::CandidType;
+use human_readable::{HumanReadablePrincipal, ToHumanReadable};
 use serde::{Deserialize, Serialize};
 use types::CanisterId;
 
@@ -13,4 +14,21 @@ pub enum Response {
     Success,
     NotFound,
     InternalError(String),
+}
+
+#[derive(Serialize)]
+pub struct HumanReadableArgs {
+    governance_canister_id: HumanReadablePrincipal,
+    delete_group: bool,
+}
+
+impl ToHumanReadable for Args {
+    type Target = HumanReadableArgs;
+
+    fn to_human_readable(&self) -> Self::Target {
+        HumanReadableArgs {
+            governance_canister_id: self.governance_canister_id.into(),
+            delete_group: self.delete_group,
+        }
+    }
 }