+++ backend/canisters/storage_index/api/src/updates/mod.rs
@@ -5,6 +5,5 @@ pub mod c2c_sync_bucket;
 pub mod remove_accessor;
 pub mod remove_user;
 pub mod set_bucket_full;
-pub mod set_governance_principals;
 pub mod update_user_id;
 pub mod upgrade_bucket_canister_wasm;