+++ backend/canisters/user_index/impl/src/updates/mod.rs
@@ -14,7 +14,6 @@ pub mod pay_for_diamond_membership;
 pub mod register_user;
 pub mod remove_platform_moderator;
 pub mod remove_platform_operator;
-pub mod set_governance_principals;
 pub mod set_max_concurrent_user_canister_upgrades;
 pub mod set_neuron_controller_for_initial_airdrop;
 pub mod set_user_upgrade_concurrency;