+++ frontend/app/src/components/landingpages/blog/Communities.svelte
@@ -3,8 +3,6 @@
     import BlogScreenshot from "./BlogScreenshot.svelte";
 </script>
 
-<h1>Communities in depth</h1>
-
 <section>
     <h2>Introduction</h2>