+++ backend/canisters/cycles_dispenser/CHANGELOG.md
@@ -5,6 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Added
+
+- Added missing proposal validation functions ([#3298](https://github.com/open-ic/open-chat/pull/3298))
+
+### Removed
+
+- Removed `set_governance_principals` ([#3301](https://github.com/open-ic/open-chat/pull/3301))
+
+## [[2.0.632](https://github.com/open-ic/open-chat/releases/tag/v2.0.632-cycles_dispenser)] - 2023-03-10
+
 ### Changed
 
 - Added `icp_burn_amount` to `update_config` args ([#3272](https://github.com/open-ic/open-chat/pull/3272))