+++ backend/canisters/local_user_index/CHANGELOG.md
@@ -9,10 +9,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Event for setting user upgrade concurrency ([#3268](https://github.com/open-ic/open-chat/pull/3268))
 - Added `OpenChatBotMessage` event type ([#3274](https://github.com/open-ic/open-chat/pull/3274))
+- Return group summary if user tries to join group they are already in ([#3296](https://github.com/open-ic/open-chat/pull/3296))
 
 ### Changed
 
 - Switch to using the new `c2c_notify_events` endpoint ([#3283](https://github.com/open-ic/open-chat/pull/3283))
+- Set upgrade concurrency to 10 ([#3302](https://github.com/open-ic/open-chat/pull/3302))
 
 ### Removed